# [SPARK-26378][SQL] Restore performance of queries against wide CSV/JSON tables