@MaxGekk 2018-12-17
May I ask you to update `CSVBenchmark-results.txt` and add your benchmarks to `CSVBenchmark` if they don't overlap with `multiColumnsBenchmark`