@bersprockets 2018-12-17
> May I ask you to update `CSVBenchmark-results.txt` and add your benchmarks to `CSVBenchmark` if they don't overlap with `multiColumnsBenchmark`

@MaxGekk Thanks for the pointer. I didn't know about these tools.