@MaxGekk 2018-12-23
@dongjoon-hyun @heary-cao Could you help to re-run CSVBenchmarks on:
```
OpenJDK 64-Bit Server VM 1.8.0_191-b12 on Linux 3.10.0-862.3.2.el7.x86_64
Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
```