@MaxGekk 2019-01-24
Please, add to the PR updated `JSONBenchmark-results.txt`.