@HyukjinKwon 2019-01-24
We don't necessarily use the same environment if the performance improvement looks distinguishing. We can check the ratio. If the last change is not reflected in the benchmark, yes, let's rerun.