@bersprockets 2019-01-24
@MaxGekk @HyukjinKwon Yes, I should rerun the benchmarks since Spark itself has changed (there have been nearly 200 commits to Spark since I last ran the benchmarks). However, currently running the JSON benchmark twice would take at least 4-5 hours (due to SPARK-26711). So I am waiting until that is fixed.