@HyukjinKwon 2019-01-28
I think it's clear that it saves the tame. Let's just push the results file here.