@bersprockets 2019-01-29
@HyukjinKwon Please note that I additionally changed JsonBenchmark.scala to run inference benchmarks with inferTimestamp=false.