@HyukjinKwon 2019-01-30
Merged to master.